# openapi_emma.json
{"openapi":"3.1.0","info":{"title":"EMMA Evolution Gatekeeper","version":"0.1.0"},"paths":{"/health":{"get":{"summary":"Health","description":"Basic health endpoint used by orchestration and local checks.\n\nReturns service status, Redis connectivity, and a quick pending proposals count.","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/terminal/session":{"post":{"summary":"Create Session","operationId":"create_session_terminal_session_post","parameters":[{"name":"x-secret-key","in":"header","required":true,"schema":{"type":"string","title":"X-Secret-Key"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/terminal/session/{session_id}/close":{"post":{"summary":"Close Session","operationId":"close_session_terminal_session__session_id__close_post","parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"string","title":"Session Id"}},{"name":"x-secret-key","in":"header","required":true,"schema":{"type":"string","title":"X-Secret-Key"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/notify":{"post":{"summary":"Receive","description":"Endpoint for Lucy (SelfâExtension Engine) to push events.\nExpected keys: `event` (str) and `payload` (dict).","operationId":"receive_notify_post","parameters":[{"name":"x-secret-key","in":"header","required":true,"schema":{"type":"string","title":"X-Secret-Key"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Event"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/skit/generate":{"post":{"summary":"Generate Skit","description":"Generate a multiâspeaker skit.\nEnforces a simple rateâlimit of 20 skits per hour using Redis.\nPublishes a `skit:generate` event for downstream processing.","operationId":"generate_skit_skit_generate_post","parameters":[{"name":"x-secret-key","in":"header","required":true,"schema":{"type":"string","title":"X-Secret-Key"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Request"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/command/{cmd}":{"post":{"summary":"Command","description":"Accept a command from the UI and forward it to Lucy via the neuroâbus.\nSupported `cmd` values: `request_explanation`, `request_risk_analysis`,\n`request_diff_summary`, `request_sandbox_logs`.","operationId":"command_command__cmd__post","parameters":[{"name":"cmd","in":"path","required":true,"schema":{"type":"string","title":"Cmd"}},{"name":"x-secret-key","in":"header","required":true,"schema":{"type":"string","title":"X-Secret-Key"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommandRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/terminal/exec":{"post":{"summary":"Terminal Exec","description":"Execute a shell command via Emma's terminal service.\n\nReturns stdout/stderr/exit_code and stores the entry in history.","operationId":"terminal_exec_terminal_exec_post","parameters":[{"name":"x-secret-key","in":"header","required":true,"schema":{"type":"string","title":"X-Secret-Key"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ShellRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/terminal/history":{"get":{"summary":"Terminal History","operationId":"terminal_history_terminal_history_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":50,"title":"Limit"}},{"name":"x-secret-key","in":"header","required":true,"schema":{"type":"string","title":"X-Secret-Key"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"CommandRequest":{"properties":{"proposal_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Proposal Id"}},"type":"object","title":"CommandRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ShellRequest":{"properties":{"command":{"type":"string","title":"Command"},"timeout":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Timeout","default":60}},"type":"object","required":["command"],"title":"ShellRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}